+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -531,7 +531,6 @@
       "hasApplyArgsData",
       "hasAuthHeaders",
       "hasDeps",
-      "hasFromBase64",
       "hasInSkipHydrationBlockFlag",
       "hasLift",
       "hasMatchingDehydratedView",
